requirements.txt
# Django и основные зависимости
Django>=5.1,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8

# Django Channels для WebSockets
channels>=4.0.0
channels-redis>=4.1.0
asgiref>=3.7.0
daphne>=4.0.0  # ASGI сервер для WebSockets

# Celery для асинхронных задач
celery>=5.3.0
redis>=5.0.0

# MySQL драйвер
#mysqlclient>=2.2.0
# Альтернатива (если mysqlclient не установится):
PyMySQL>=1.1.0

# Telegram API клиенты
hydrogram>=0.2.0  # Для личных аккаунтов (MTProto)
requests>=2.31.0  # Для Bot API
aiohttp>=3.9.0  # Для асинхронных HTTP запросов

# Утилиты
python-dateutil>=2.8.2

# Разработка (опционально)
ipython>=8.12.0
django-extensions>=3.2.0

# Тестирование
websocket-client>=1.6.0
